feat(webui): add dedicated task route and remove task list view
Introduce a tanstack /task route that renders ChatPage via LiveStoreDefaultProvider and update webview entry points to link to it, replacing the old paginated tasks list implementation.
